bug: performance issue, 18sec til fcp, mantine is one cause. refactoring to eliminate passing createStyles props to child component
