+++ blogcms-react/src/pages/AuthPage.tsx
@@ -11,14 +11,14 @@ import {
   } from '@mantine/core';
 import { useForm } from '@mantine/form';
 import { UserCredentials } from '@supabase/supabase-js';
+import { useState } from 'react';
+import LoginForm from '../components/LoginForm';
+import RegisterForm from '../components/RegisterForm';
 import { sbClient } from '../services/sb';
 /**
  * Interfaces and types
  */
-interface LoginValues {
-    email: string
-    password: string
-}
+
   
   const useStyles = createStyles((theme) => ({
     wrapper: {
@@ -56,45 +56,15 @@ interface LoginValues {
   }));
   
   export function AuthPage() {
+    const [registerPage, setRegisterPage] = useState<boolean>(false)
     const { classes } = useStyles();
-    const form = useForm({
-        initialValues: {
-            email: '',
-            password: ''
-        },
-        validate: {
-            email: (value) => (/^\S+@\S+$/.test(value) ? null : 'Invalid email'),
-          },
-    })
 
-  async function handleSubmit (values: LoginValues) {
-    console.log(values.email, values.password)
-    sbClient.auth.signIn(values as UserCredentials)
-  } 
+
     return (
       <div className={classes.wrapper}>
-        <form onSubmit={form.onSubmit((values) => handleSubmit(values))}>
         <Paper className={classes.form} radius={0} p={30}>
-          <Title order={2} className={classes.title} align="center" mt="md" mb={50}>
-            Welcome to BlogCMS
-          </Title>
-  
-          <TextInput label="Email address" placeholder="hello@gmail.com" size="md"
-          required
-          {...form.getInputProps('email')} />
-          <PasswordInput label="Password" placeholder="Your password" mt="md" size="md"
-          required
-          {...form.getInputProps('password')} />
-          <Button type="submit" fullWidth mt="xl" size="md">
-            Login
-          </Button>
-  
-          <Text align="center" mt="md">
-            Contact your administrator to request credentials.
-            
-          </Text>
+          {registerPage === false? <LoginForm classes={classes}/> : <RegisterForm classes={classes} />}
         </Paper>
-        </form>
       </div>
     );
   }
